add rpc create cart

+++ gapi/converter.go
@@ -23,8 +23,8 @@ func convertProduct(product db.Product) *pb.Product {
 	return &pb.Product{
 		Name:          product.Name,
 		Description:   product.Description,
-		Price:         product.Price,
-		StockQuantity: product.StockQuantity,
+		Price:         int32(product.Price),
+		StockQuantity: int32(product.StockQuantity),
 		CreatedAt:     timestamppb.New(product.CreatedAt),
 	}
 }
